Fix spell check false positives by ignoring words
The codespell spellchecker tool is used to automatically detect commonly misspelled words in the files of this project.

The misspelled words dictionary was expanded in the latest release of codespell. Some of the text in the project
codebase happens to match against newly added entries, which caused codespell to produce false misspelled word
detections.

Since the code that produced the detections is correct and intended, the false positives are resolved by configuring
codespell to ignore the problematic word.